Move type guard functions to new file
